requirements.txt
Flask==2.2.3
json5==0.9.11
matplotlib==3.7.1
matplotlib-inline==0.1.6
mime==0.1.0
nltk==3.8.1
num2words==0.5.12
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.2
openai-secret-manager==0.0.2
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
QtPy==2.3.1
random2==1.0.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
SpeechRecognition==3.10.0
sympy==1.11.1
twilio==8.1.0
webencodings==0.5.1

